Cancel lot changes

+++ src/main/java/com/sericulture/marketandauction/model/entity/ReelerAuction.java
@@ -31,7 +31,7 @@ public class ReelerAuction extends BaseEntity implements Serializable {
     private int allottedLotId;
 
     @Column(name = "REELER_ID")
-    private BigInteger reelerId;
+    private int reelerId;
 
     @Column(name = "AMOUNT")
     private int amount;
